+++ CHANGELOG.md
@@ -7,7 +7,8 @@ Changes in this release:
 - Added extensions mechanism (#565, #1185)
 - Add /serverstatus api call to get version info, loaded slices and extensions (#1184)
 - Support to set environment variables on the Inmanta server and its agents
-- Split of server recompile into separate extension (#1183)
+- Split of server recompile into separate server slice (#1183)
+- Add API to inspect compiler service queue (#1252)
 - Define explicit path in protocol methods
 - Added support for schema management for multiple slices in the same database (#1207)
 - Marked pypi package as typed